we are using a comparison with sizeof(long) to figure the size of time_t this is unfortunate since on freebsd in 64bit mode time_t is 32 bit while long is 64 bit
git-svn-id: svn://svn.oetiker.ch/rrdtool/branches/1.3/program@1997 a5681a0c-68f1-0310-ab6d-d61299d08faa

+++ src/rrd_restore.c
@@ -960,7 +960,7 @@ static int parse_tag_rrd(
             status = get_ulong_from_node(doc, child,
                                         &rrd->stat_head->pdp_step);
         else if (xmlStrcmp(child->name, (const xmlChar *) "lastupdate") == 0) {
-            if (sizeof(time_t) == sizeof(long)) {
+            if (sizeof(time_t) == sizeof(int)) {
                status = get_long_from_node(doc, child, (long *)&rrd->live_head->last_up);
             }
             else { if (sizeof(time_t) == sizeof(long long)) {
